Improved match context accuracy

+++ source/CycleDetection/CycleDetection.cs
@@ -7,6 +7,9 @@ using AxiomProfiler.Utilities;
 
 namespace AxiomProfiler.CycleDetection
 {
+    using ConstraintElementType = Tuple<Term, int>;
+    using ConstraintType = List<Tuple<Term, int>>;
+
     public class CycleDetection
     {
         private readonly List<Instantiation> path; // Path on which to search for a cycle.
@@ -521,7 +524,7 @@ namespace AxiomProfiler.CycleDetection
                         var newBody = GeneralizeAtBindings(inst.concreteBody, inst.Quant.BodyTerm.Args.Last(), nextBindingInfo, out var usedConcrete);
                         if (newBody != null)
                         {
-                            if (afterNextBindingInfo != null) UpdateBindingsForReplacementTerm(usedConcrete, newBody, afterNextBindingInfo, new List<Term>(), new List<Term>());
+                            if (afterNextBindingInfo != null) UpdateBindingsForReplacementTerm(usedConcrete, newBody, afterNextBindingInfo, new ConstraintType(), new ConstraintType());
                             inst.concreteBody = newBody;
                         }
                         else
@@ -566,7 +569,7 @@ namespace AxiomProfiler.CycleDetection
                 {
                     if (!lastBindingInfo.bindings.ContainsKey(key))
                     {
-                        lastBindingInfo.bindings[key] = default(Tuple<List<List<Term>>, Term>);
+                        lastBindingInfo.bindings[key] = default(Tuple<List<ConstraintType>, Term>);
                     }
                 }
             }
@@ -1997,9 +2000,10 @@ namespace AxiomProfiler.CycleDetection
             {
                 for (var i = 0; i < constraint.Count; ++i)
                 {
-                    if (constraint[i].id == t.id)
+                    var el = constraint[i];
+                    if (el.Item1.id == t.id)
                     {
-                        constraint[i] = resultTerm;
+                        constraint[i] = Tuple.Create(resultTerm, el.Item2);
                         break;
                     }
                 }
@@ -2591,7 +2595,7 @@ namespace AxiomProfiler.CycleDetection
         }
 
         // Called after GeneralizeAtBindings() to obtain bindings for a match based on the modified quantifier yield.
-        private static void UpdateBindingsForReplacementTerm(Term originalTerm, Term newTerm, BindingInfo bindingInfo, List<Term> history, List<Term> newHistory)
+        private static void UpdateBindingsForReplacementTerm(Term originalTerm, Term newTerm, BindingInfo bindingInfo, ConstraintType history, ConstraintType newHistory)
         {
 
             // Update bindings
@@ -2625,16 +2629,14 @@ namespace AxiomProfiler.CycleDetection
             }
 
             if (newTerm.generalizationCounter < 0) {
-                history.Add(originalTerm);
-                newHistory.Add(newTerm);
-
-                foreach (var argTuple in originalTerm.Args.Zip(newTerm.Args, Tuple.Create))
+                for (var i = 0; i < Math.Min(originalTerm.Args.Length, newTerm.Args.Length); ++i)
                 {
-                    UpdateBindingsForReplacementTerm(argTuple.Item1, argTuple.Item2, bindingInfo, history, newHistory);
+                    history.Add(Tuple.Create(originalTerm, i));
+                    newHistory.Add(Tuple.Create(newTerm, i));
+                    UpdateBindingsForReplacementTerm(originalTerm.Args[i], newTerm.Args[i], bindingInfo, history, newHistory);
+                    history.RemoveAt(history.Count - 1);
+                    newHistory.RemoveAt(newHistory.Count - 1);
                 }
-
-                history.RemoveAt(history.Count - 1);
-                newHistory.RemoveAt(newHistory.Count - 1);
             }
         }
 
@@ -2671,15 +2673,16 @@ namespace AxiomProfiler.CycleDetection
             // 4th and 5th element are id and generalization counter (kept if all terms agree)
             // last element is a T term all terms are equal to or null
             var candidates = new Dictionary<string, Tuple<int, string, int, int, int, Term>>();
-            var concreteHistories = terms.Select(_ => new Stack<Term>()).ToArray();
-            var generalizedHistory = new Stack<Term>();
+            var concreteHistories = terms.Select(_ => new Stack<ConstraintElementType>()).ToArray();
+            var generalizedHistory = new Stack<ConstraintElementType>();
 
             while (true)
             {
                 // check for backtrack condition
+                Term mostRecent = null;
                 if (generalizedHistory.Count > 0)
                 {
-                    var mostRecent = generalizedHistory.Peek();
+                    mostRecent = generalizedHistory.Peek().Item1;
                     if (mostRecent.Args.Length == 0 || mostRecent.Args[0] != null)
                     {
                         // this term is full --> backtrack
@@ -2702,13 +2705,32 @@ namespace AxiomProfiler.CycleDetection
                     }
                 }
 
+                if (mostRecent != null)
+                {
+                    // Update the histories with the argument index for this child
+                    var argIdx = Array.FindLastIndex(mostRecent.Args, t => t == null);
+                    var top = generalizedHistory.Peek();
+                    generalizedHistory.Pop();
+                    generalizedHistory.Push(Tuple.Create(top.Item1, argIdx));
+                    foreach (var history in concreteHistories)
+                    {
+                        top = history.Peek();
+                        history.Pop();
+                        history.Push(Tuple.Create(top.Item1, argIdx));
+                    }
+                }
+
                 // find candidates for the next term.
                 int i = 0;
-                foreach (var currentTermAndBindings in todoStacks.Select(stack => stack.Peek()).Zip(highlightInfoInsts.Skip(blameWrapAround ? 1 : 0), Tuple.Create))
+                foreach (var currentTermAndBindings in todoStacks.Select(stack => stack.Peek())
+                    .Zip(highlightInfoInsts.Skip(blameWrapAround ? 1 : 0), Tuple.Create)
+                    .Zip(concreteHistories, (t, h) => Tuple.Create(t.Item1, t.Item2, h)))
                 {
                     var currentTerm = currentTermAndBindings.Item1;
                     var boundIn = currentTermAndBindings.Item2;
-                    var boundTo = boundIn.bindingInfo.bindings.Keys.FirstOrDefault(k => boundIn.bindingInfo.bindings[k].Item2.id == currentTerm.id);
+                    var history = currentTermAndBindings.Item3.Reverse().ToList();
+                    var boundTo = boundIn.bindingInfo.bindings.Keys.FirstOrDefault(k => boundIn.bindingInfo.bindings[k].Item2.id == currentTerm.id &&
+                        constraintsSat(history, boundIn.bindingInfo.bindings[k].Item1));
                     if (boundTo != null && generalizedBindingInfo.bindings.TryGetValue(boundTo, out var existing) && existing.Item2.id < -1)
                     {
                         /* The term was bound to a quantified variable and was already encountered and generalized in this term,
@@ -2735,7 +2757,7 @@ namespace AxiomProfiler.CycleDetection
                     {
                         if (!generalizedBindingInfo.bindings.TryGetValue(bindingKey, out var existing))
                         {
-                            existing = Tuple.Create(new List<List<Term>>(), currTerm);
+                            existing = Tuple.Create(new List<ConstraintType>(), currTerm);
                             generalizedBindingInfo.bindings[bindingKey] = existing;
                         }
                         else if (existing.Item2.id != currTerm.id)
@@ -2751,7 +2773,7 @@ namespace AxiomProfiler.CycleDetection
                     {
                         if (!generalizedBindingInfo.equalities.TryGetValue(key, out var eqList))
                         {
-                            eqList = new List<Tuple<List<Term>, Term>>();
+                            eqList = new List<Tuple<ConstraintType, Term>>();
                             generalizedBindingInfo.equalities[key] = eqList;
                         }
                         eqList.Add(Tuple.Create(generalizedHistory.Reverse().ToList(), currTerm));
@@ -2759,8 +2781,9 @@ namespace AxiomProfiler.CycleDetection
                 }
 
                 // always push the generalized term, because it is one term 'behind' the others
-                generalizedHistory.Push(currTerm);
-                foreach (var pair in concreteHistories.Zip(todoStacks, Tuple.Create)) pair.Item1.Push(pair.Item2.Peek());
+                
+                generalizedHistory.Push(Tuple.Create(currTerm, -1));
+                foreach (var pair in concreteHistories.Zip(todoStacks, Tuple.Create)) pair.Item1.Push(Tuple.Create(pair.Item2.Peek(), -1));
                 // push children if applicable
                 if (currTerm.Args.Length > 0 && currTerm.generalizationCounter < 0)
                 {
@@ -2772,7 +2795,7 @@ namespace AxiomProfiler.CycleDetection
             }
         }
 
-        private bool isBlameTerm(List<Instantiation> childInsts, Stack<Term>[] todoStacks, Stack<Term>[] concreteHistories, bool flipped, out List<Term> boundTo)
+        private bool isBlameTerm(List<Instantiation> childInsts, Stack<Term>[] todoStacks, Stack<ConstraintElementType>[] concreteHistories, bool flipped, out List<Term> boundTo)
         {
             boundTo = null;
 
@@ -2796,7 +2819,7 @@ namespace AxiomProfiler.CycleDetection
             return true;
         }
 
-        private bool isEqTerm(List<Instantiation> childInsts, Stack<Term>[] todoStacks, Stack<Term>[] concreteHistories, bool flipped, out List<Term> boundTos)
+        private bool isEqTerm(List<Instantiation> childInsts, Stack<Term>[] todoStacks, Stack<ConstraintElementType>[] concreteHistories, bool flipped, out List<Term> boundTos)
         {
             //We try to find an equality that was used in the concrete case and assume that it should also be extended to the generalized case
             var eqsPerIteration = todoStacks.Select(s => s.Peek()).Zip(concreteHistories, Tuple.Create).Zip(childInsts.Skip(flipped ? 1 : 0), (checkTermAndHistory, checkInst) =>
@@ -2834,7 +2857,7 @@ namespace AxiomProfiler.CycleDetection
                         var bindingInfo = iterationResult.Item3.bindingInfo;
                         if (!bindingInfo.equalities.TryGetValue(boundTo, out var equalityLhs))
                         {
-                            equalityLhs = new List<Tuple<List<Term>, Term>>();
+                            equalityLhs = new List<Tuple<ConstraintType, Term>>();
                             bindingInfo.equalities[boundTo] = equalityLhs;
                         }
                         // Since the concrete term previously occured under some constraints and now becomes a distinct blame term,
@@ -2855,31 +2878,31 @@ namespace AxiomProfiler.CycleDetection
             return boundTos.Any();
         }
 
-        private static void RemovePrefixFromConstraints(List<Term> prefix, BindingInfo bindingInfo)
+        private static void RemovePrefixFromConstraints(ConstraintType prefix, BindingInfo bindingInfo)
         {
             var allMatchContexts = bindingInfo.bindings.SelectMany(kv => kv.Value.Item1)
                 .Concat(bindingInfo.equalities.SelectMany(kv => kv.Value.Select(t => t.Item1)));
             foreach (var context in allMatchContexts)
             {
-                if (context.Count >= prefix.Count && prefix.Zip(context, (t1, t2) => t1.id == t2.id).All(x => x))
+                if (context.Count >= prefix.Count && prefix.Zip(context, (t1, t2) => t1.Item1.id == t2.Item1.id && t1.Item2 == t2.Item2).All(x => x))
                 {
                     context.RemoveRange(0, prefix.Count);
                 }
             }
         }
 
-        private static bool constraintsSat(List<Term> generalizedHistory, List<Term> constraint)
+        private static bool constraintsSat(ConstraintType generalizedHistory, ConstraintType constraint)
         {
             return constraintsSat(generalizedHistory, Enumerable.Repeat(constraint, 1));
         }
 
-        private static bool constraintsSat(List<Term> gerneralizeHistory, IEnumerable<List<Term>> constraints)
+        private static bool constraintsSat(ConstraintType gerneralizeHistory, IEnumerable<ConstraintType> constraints)
         {
             if (!constraints.Any()) return true;
             return (from constraint in constraints
                     where constraint.Count <= gerneralizeHistory.Count
                     let slice = gerneralizeHistory.GetRange(gerneralizeHistory.Count - constraint.Count, constraint.Count)
-                    select slice.Zip(constraint, (term1, term2) => term1.id == term2.id))
+                    select slice.Zip(constraint, (term1, term2) => term1.Item1.id == term2.Item1.id && term1.Item2 == term2.Item2))
                                 .Any(intermediate => intermediate.All(val => val));
         }
 
@@ -2922,7 +2945,7 @@ namespace AxiomProfiler.CycleDetection
             return iterationNumberings == null ? idx : iterationNumberings[idx];
         }
 
-        private Term getGeneralizedTerm(Dictionary<string, Tuple<int, string, int, int, int, Term>> candidates, Stack<Term>[] todoStacks, Stack<Term> generalizedHistory, List<int> instantiationNumberings, bool overrideReplacements)
+        private Term getGeneralizedTerm(Dictionary<string, Tuple<int, string, int, int, int, Term>> candidates, Stack<Term>[] todoStacks, Stack<ConstraintElementType> generalizedHistory, List<int> instantiationNumberings, bool overrideReplacements)
         {
             Term currTerm;
             if (candidates.Count == 1)
@@ -2994,9 +3017,9 @@ namespace AxiomProfiler.CycleDetection
         /// <summary>
         /// Adds the specified term to the arguments of its parent term.
         /// </summary>
-        private static void addToGeneralizedTerm(Stack<Term> generalizedHistory, Term currTerm)
+        private static void addToGeneralizedTerm(Stack<ConstraintElementType> generalizedHistory, Term currTerm)
         {
-            var genParent = generalizedHistory.Count > 0 ? generalizedHistory.Peek() : null;
+            var genParent = generalizedHistory.Count > 0 ? generalizedHistory.Peek().Item1 : null;
             // connect to parent
             if (genParent != null)
             {
@@ -3040,7 +3063,7 @@ namespace AxiomProfiler.CycleDetection
             var key = currentTerm.Name + currentTerm.GenericType + currentTerm.Args.Length + "_" + currentTerm.generalizationCounter;
 
             var existingReplacements = bindingInfo.bindings.Where(kv => kv.Value.Item2.id == currentTerm.id)
-                .SelectMany(kv => bindingInfo.equalities.TryGetValue(kv.Key, out var eqs) ? eqs : Enumerable.Empty<Tuple<List<Term>, Term>>())
+                .SelectMany(kv => bindingInfo.equalities.TryGetValue(kv.Key, out var eqs) ? eqs : Enumerable.Empty<Tuple<ConstraintType, Term>>())
                 .SelectMany(t => replacementDict.TryGetValue(Tuple.Create(iteration, t.Item2.id), out var gen) ? gen : nonGenTerm);
             //TODO: keep tracking all existing replacements
             var generalization = existingReplacements.FirstOrDefault();
@@ -3118,14 +3141,14 @@ namespace AxiomProfiler.CycleDetection
             }
             foreach (var termAndMatchContext in bindingInfo.bindings.Values)
             {
-                if (termAndMatchContext != default(Tuple<List<List<Term>>, Term>))
+                if (termAndMatchContext != default(Tuple<List<ConstraintType>, Term>))
                 {
                     termAndMatchContext.Item2.highlightTemporarily(format, PrintConstants.blameColor, termAndMatchContext.Item1);
                 }
             }
             foreach (var termAndMatchContext in bindingInfo.bindings.Where(kv1 => kv1.Key.id == -1).Select(kv2 => kv2.Value))
             {
-                if (termAndMatchContext != default(Tuple<List<List<Term>>, Term>))
+                if (termAndMatchContext != default(Tuple<List<ConstraintType>, Term>))
                 {
                     termAndMatchContext.Item2.highlightTemporarily(format, PrintConstants.bindColor, termAndMatchContext.Item1);
                 }
